requirements.txt
PyMuPDF==1.25.3
Pillow==11.1.0
pytesseract==0.3.13
